requirements/development.txt
-r base.txt

django-nose==1.4.6